Created a new build system
With these changes, we can now run tasks from the root directory in vscode.

Added the .vscode folder which creates the same task commands we all love, but added a prompt and some extra commands to each task to automatically move to the desired directory for running the task, based on the prompt (a drop down menu). To add a new folder to the potential build list, you need to manually go into the tasks.json file and add it to the input variable in the objects list. There is a comment explaining it.